Fixes to ensure backup camera shutoff

+++ source/dashcam_logger.cpp
@@ -263,13 +263,14 @@ void backupCameraToggleHandler(int signumber, siginfo_t *siginfo, void *pointer)
   }
   else {
     //sleep(5);   // fulfill FMVSS by waiting 5 sec to kill backup camera after shifting out of reverse
-    //system("killall backup_cam_helper");
-    /*if(bcamerapid > 1) {
+    system("killall backup_cam_helper");
+    if(bcamerapid > 1) {
       kill(bcamerapid,SIGKILL);
       waitpid(bcamerapid, &status, -1);
     }
     bcamerapid = -5;
-    system("pkill -f port=9003");*/
+    system("killall -SIGINT gst-launch-1.0");
+    //system("pkill -f port=9003");
   }
 }
